Allow admin(s) to add, remove, or edit available tutors and other admins. (#58)
* refactored MainActivity

* refactored admin activity

* refactored student activity

* refactor new student activity

* refactored sign out fragment

* refactored tutor activity

* put db related classes in database folder

* added polymorphism on User to reduce coupling

* codestyle changes

* fixed tests

* related to functional requirement 6 #12

* related to #12

* Fixes #12

* Added Comments and KotlinDocumentation for AddUserActivity, AdminActivity, AdminListActivity, User, UserDao, EditUserActivity

* AdminListActivity gets data based on userLevel now (#55)

The list of information that the AdminListActivity shows is now based on the parameter "level" passed as one of the extra values in the intent. This allows for a more general purpose implementation of presenting the applicable database information.

* can format of xml file

* changed format for xml file

* Checked out master files for cleaner pull request

Majority of the files are modified to account for minor changes in
miscellaneous file changes.

* fixed editUserActivity to update the user db and send only the id back. Commented 1 line to make sure we have a fallback.

* fixed "modify the user class and add a method that will return this intent to you".
function setIntent has second param asking for Class<EditUserActivity>. Not good for other classes as of right now
left line commented in AdminListActivity in case of fallback

* Renamed to better match Activity functionality

Instead of the misleading name of "AdminListActivity", which can be
confusing and lead people to think that the activity is strictly only
for modifying other admins, the activity is now called
"AdminModifyUsersList". The name now indicated that depending on the
users, the activity will allow one to modify the list of applicable
users.